LPD-22034 Move handleActionClick function to utils
