+++ modules/apps/frontend-data-set/frontend-data-set-web/types/src/main/resources/META-INF/resources/utils/actionItems/handleActionClick.d.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {IItemsActions} from '../index';
+import {IItemsActions} from '../../index';
 declare const handleActionClick: ({
 	action,
 	closeMenu,
